In this example, the program loads five audio files (intro, verse1, chorus, verse2 and outro) and concatenates them together to form a complete song. The final song is then exported as an mp3 file.
You can replace the path of the audio files and the format of the song with the ones you want. Also you can use fade_in() and fade_out() functions to fade in/out the audio files and make smooth transitions between them.

This is a simple example that can be customized to suit your needs, such as mixing multiple tracks together, adjusting the volume of each track, adding effects, etc. Pydub library offers a lot of functionalities to work with audio files and you can use them to achieve your desired results.